% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggcorplot}
\alias{ggcorplot}
\title{Correlation plot}
\usage{
ggcorplot(rpcaObj, pcs = c(1, 2))
}
\arguments{
\item{rpcaObj}{object containing the \code{sdev} component, such as that returned
by \code{rpca}}

\item{pcs}{array_like \cr
an array with two values indicating which two PCs should be plotted,
by default the first two PCs are used, e.g., \eqn{c(1,2)}. \cr}

\item{.................}{.}
}
\description{
Creates a pretty plot which is showing the correlation of
   the original variable with the principal component (PCs).
}
\examples{
#

}
\seealso{
\code{\link{rpca}}, \code{\link[ggplot2]{ggplot}}
}

