% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggscreeplot}
\alias{ggscreeplot}
\title{Pretty Screeplot}
\usage{
ggscreeplot(rpcaObj, type = c("var", "ratio", "cum", "eigenvals"))
}
\arguments{
\item{type}{str c('var', 'ratio', 'cum', 'eigenvals'), optional \cr}

\item{x}{object containing the \code{sdev} component, such as that returned
by \code{rpca}}

\item{.................}{.}
}
\description{
Creates a pretty screeplpot using \code{ggplot2}. By default the explained variance is plotted
agaings the number of the principal component.
Alternatively the eigenvalues, explained variance ratio or the cumulative
explained variance ratio can be plotted.
}
\examples{
#


}
\seealso{
\code{\link{rsvd}}, \code{ggplot2}
}

