% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highway.R
\docType{data}
\name{highway}
\alias{highway}
\title{Highway}
\format{An object of class \code{"rsvd"}.}
\source{
\href{http://changedetection.net/}{changedetection.net}
}
\usage{
data(highway)
}
\description{
176x144 grayscaled (8 bit [0-255]) surveillance video with 200 frames.
Each frame is flattened, and stored as a column vector.
}
\examples{
library(rsvd)
data(highway)

# Reshape and display the 100th frame:
frame <- matrix(highway[,100], ncol=144, nrow=176)
image(frame, col = gray((0:255)/255))

}
\references{
N. Goyette, P.-M. Jodoin, F. Porikli, J. Konrad, and P. Ishwar,
           changedetection.net: A new change detection benchmark dataset,
           in Proc. IEEE Workshop on Change Detection (CDW-2012) at CVPR-2012,
           Providence, RI, 16-21 Jun., 2012.
}
\keyword{video}

