% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rid.R
\name{rid}
\alias{rid}
\title{Randomized interpolative decomposition (ID).}
\usage{
rid(A, k = NULL, mode = "column", p = 10, q = 0, idx_only = FALSE,
  rand = "TRUE")
}
\arguments{
\item{A}{Array_like. \cr
A numeric input matrix (or data frame), with dimensions \eqn{(m, n)}. \cr
If the data contain \eqn{NA}s na.omit is applied.}

\item{k}{Int, optional. \cr
Determines the number of rows/columns to be selected. 
It is required that \eqn{k} is smaller or equal to \eqn{min(m,n)}.}

\item{mode}{String c('column', 'row'). \cr
Determines whether a subset of columns or rows is selected.}

\item{p}{Int, optional. \cr
Oversampling parameter (default \eqn{p=10}).}

\item{q}{Int, optional. \cr
Number of power iterations (default \eqn{q=0}).}

\item{idx_only}{Bool (\eqn{TRUE}, \eqn{FALSE}), optional. \cr
If (\eqn{TRUE}), the index set \code{idx} is returned, but not the matrix \code{C} or \code{R}. 
This is more memory efficient, when dealing with large-scale data.}

\item{rand}{Bool (\eqn{TRUE}, \eqn{FALSE}). \cr
If (\eqn{TRUE}), a probabilistic strategy is used, otherwise a deterministic algorithm is used.}

\item{.................}{.}
}
\value{
\code{id} returns a list with class \eqn{id} containing the following components:
   \item{C}{ Array_like. \cr
             Column subset \eqn{C = A[,idx]}, if \code{mode='column'}; array with dimensions \eqn{(m, k)}.
   }

   \item{R}{ Array_like. \cr
              Row subset \eqn{R = A[idx, ]}, if \code{mode='row'}; array with dimensions \eqn{(k, n)}.
   }
   
   \item{Z}{ Array_like \cr
            Well conditioned matrix; Dependng on the slected mode, this is an 
            array with dimensions \eqn{(k,n)} or \eqn{(m,k)}.
   }
   
   \item{idx}{ Array_like \cr
               The index set of the \eqn{k} selcted columns or rows used to form \eqn{C} or \eqn{R}. 
   }   
  
   \item{pivot}{ Array_like \cr
               Information on the pivoting strategy used during the decomposition. 
   }  
     
   \item{scores}{ Array_like .\cr
                  The scores (importancies) of the columns or rows of the input matrix \eqn{A}.
   } 
                  
   \item{scores.idx}{ Array_like .\cr
                  The scores (importancies) of the \eqn{k} selected columns or rows in \eqn{C} or \eqn{R}.                    
   }
   \item{.................}{.}
}
\description{
Randomized interpolative matrix decomposition.
}
\details{
Algorithm for computing the ID of a rectangular \eqn{(m, n)} matrix \eqn{A}, with target rank 
\eqn{k << min(m,n)}. The input matrix is factored as \eqn{A = C * Z}, 
using the column pivoted QR decomposition. The factor matrix \eqn{C} is formed as a subset of 
columns of \eqn{A}, also called the partial column skeleton. 
If \code{mode='row'}, then the input matrix is factored as \eqn{A = Z * R}, using the 
row pivoted QR decomposition. The factor matrix \eqn{R} is now formed as
a subset of rows of \eqn{A}, also called the partial row skeleton. 
The factor matrix \eqn{Z} contains a \eqn{(k, k)} identity matrix as a submatrix, 
and is well-conditioned. 

If \eqn{rand='TRUE'} a probabilistic strategy is used to compute the decomposition, otherwise a
deterministic algorithm is used.
}
\seealso{
\code{\link{rcur}},
}
\author{
N. Benjamin Erichson, \email{erichson@uw.edu}
}
