% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqb.R
\name{rqb}
\alias{rqb}
\title{Randomized QB Decomposition (rqb).}
\usage{
rqb(A, k = NULL, p = 10, q = 2, sdist = "normal", rand = TRUE)
}
\arguments{
\item{A}{Array_like. \cr
A real/complex input matrix (or data frame), with dimensions \eqn{(m, n)}.}

\item{k}{Int, optional. \cr
Determines the target rank of the low-rank decomposition. It should satisfy \eqn{k << min(m,n)}.}

\item{p}{Int, optional. \cr
Oversampling parameter for (default \eqn{p=10}).}

\item{q}{Int, optional. \cr
Number of power iterations (default \eqn{q=2}).}

\item{sdist}{String \eqn{c( 'unif', 'normal', 'rademacher')}, optional. \cr
Specifies the sampling distribution. \cr
\eqn{'unif'} :  Uniform `[-1,1]`. \cr
\eqn{'normal}' (default) : Normal `~N(0,1)`. \cr
\eqn{'rademacher'} : Rademacher random variates. \cr}

\item{rand}{Bool (\eqn{TRUE}, \eqn{FALSE}). \cr
If (\eqn{TRUE}), a probabilistic strategy is used, otherwise a deterministic algorithm is used.}

\item{.............}{.}
}
\value{
\code{rqb} returns a list containing the following three components:
\item{Q}{  Array_like. \cr
          Orthogonal matrix; array with dimensions \eqn{(m, k)}.
}

\item{B}{  Array_like. \cr
          Smaller matrix; array with dimensions \eqn{(k, n)}.
}
\item{.............}{.}
}
\description{
Compute the near-optimal QB decompositon of a rectangular matrix.
}
\details{
The randomized QB decompositon factors a rectangular \eqn{(m,n)} matrix \eqn{A} as
\eqn{A = Q * B}. \eqn{Q} is an \eqn{(m,k)} matrix with orthogonal columns, and \eqn{B} a \eqn{(k,n)} matrix.
The target rank is assumend to be \eqn{k << min(m,n)}.   

\eqn{p} is an oversampling parameter to improve the approximation.
A value between 5 and 10 is recommended, and \eqn{p=10} is set by default.

The parameter \eqn{q} specifies the number of power (subspace) iterations
to reduce the approximation error. This is recommended
if the the singular values decay slowly. In practice 1 or 2 iterations
achieve good results, however, computing power iterations increases the
computational time. The number of power iterations is set to \eqn{q=2} by default.
}
\references{
\itemize{
  \item  [1] N. Halko, P. Martinsson, and J. Tropp.
         "Finding structure with randomness: probabilistic
         algorithms for constructing approximate matrix
         decompositions" (2009).
         (available at arXiv \url{http://arxiv.org/abs/0909.4061}).
  \item  [2] S. Voronin and P.Martinsson.
         "RSVDPACK: Subroutines for computing partial singular value
         decompositions via randomized sampling on single core, multi core,
         and GPU architectures" (2015).
         (available at `arXiv \url{http://arxiv.org/abs/1502.05366}).
}
}
\seealso{
\code{\link{svd}}
}
\author{
N. Benjamin Erichson, \email{erichson@uw.edu}
}
