% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrpca.R
\name{rrpca}
\alias{rrpca}
\title{Randomized robust principal component analysis (rRPCA).}
\usage{
rrpca(A, k = NULL, lamb = NULL, gamma = 1.25, rho = 1.5, maxiter = 10,
  tol = 0.001, svdalg = "auto", p = 10, q = 1, trace = FALSE, ...)
}
\arguments{
\item{A}{array_like \cr
a numeric input matrix (or data frame), with dimensions \eqn{(m, n)}. \cr
If the data contain \eqn{NA}s na.omit is applied.}

\item{k}{int, optional \cr
determines the number of principle components to compute. It is required that \eqn{k} is smaller or equal to
\eqn{n}, but it is recommended that \eqn{k << min(m,n)}.}

\item{lamb}{real, optional \cr
tuning paramter (default \eqn{gamma=m^-0.5}).}

\item{gamma}{real, optional \cr
tuning paramter (default \eqn{gamma=1.25}).}

\item{rho}{real, optional \cr
tuning paramter (default \eqn{rho=1.5}).}

\item{maxiter}{int, optional \cr
determines the maximal numbers of iterations.}

\item{tol}{real, optional \cr
tolarance paramter for the desired convergence of the algorithm.}

\item{svdalg}{str c('auto', 'rsvd', 'svd'), optional \cr
Determines which algorithm should be used for computing the singular value decomposition.
By default 'auto' is used, which decides whether to use \code{\link{rsvd}} or \code{\link{svd}},
depending on the number of principle components. If \eqn{k < min(n,m)/1.5} randomized svd is used.}

\item{p}{int, optional \cr
oversampling parameter for \eqn{rsvd}  (default \eqn{p=5}), see \code{\link{rsvd}}.}

\item{q}{int, optional \cr
number of power iterations  for \eqn{rsvd} (default \eqn{q=2}), see \code{\link{rsvd}}.}

\item{trace}{bool, optional \cr
print progress.}

\item{...}{arguments passed to or from other methods, see \code{\link{rsvd}}.}

\item{.................}{.}
}
\value{
\code{rpca} returns a list with class \eqn{rpca} containing the following components:
   \item{L}{  array_like \cr
             Low-rank component, array of shape \eqn{(m, n)}.
   }
   \item{S}{  array_like \cr
              Sparse component, array of shape \eqn{(m, n)}.
   }

   \item{k}{  int \cr
              target-rank used for the final iteration.
   }

   \item{err}{  vector \cr
              Frobenious error archieved by each iteration.
   }

   \item{.................}{.}
}
\description{
Robust principal components analysis using randomized singular value decomposition.
}
\details{
Robust principal component analysis is via PCP using the IALM (Lin et. al) algorithm.
}
\note{
...
}
\examples{

library(rsvd)
data(highway)

# Foreground/Background separation
out <- rrpca(highway, k=1, p=0, q=0, maxiter=20, svdalg='rsvd')

# Display results for the 200th video frame
par(mfrow=c(1,3))
image(matrix(highway[,200], ncol=144, nrow=176), col = gray((0:255)/255))
image(matrix(out$L[,200], ncol=144, nrow=176), col = gray((0:255)/255))
image(matrix(out$S[,200], ncol=144, nrow=176), col = gray((0:255)/255))

}
\author{
N. Benjamin Erichson, \email{nbe@st-andrews.ac.uk}
}

