% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\docType{data}
\name{tiger}
\alias{tiger}
\title{Tiger}
\format{
An object of class \code{\link[rsvd]{rsvd}}.
}
\source{
\href{https://en.wikipedia.org/wiki/File:Siberischer_tiger_de_edit02.jpg}{Wikimedia}
}
\usage{
data('tiger')
}
\description{
1600x1200 grayscaled (8 bit [0-255]/255) image.
}
\examples{
\dontrun{
library('rsvd')
data('tiger')

#Display image
image(tiger, col = gray((0:255)/255))
}
}
\references{
S. Taheri (2006). "Panthera tigris altaica", (Online image)
}
\keyword{image}
