% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscreeplot.R
\name{ggscreeplot}
\alias{ggscreeplot}
\title{Pretty Screeplot}
\usage{
ggscreeplot(rpcaObj, type = c("var", "ratio", "cum", "eigenvals"))
}
\arguments{
\item{rpcaObj}{Object returned by the \code{\link[rsvd]{rpca}} function.}

\item{type}{String c('var', 'ratio', 'cum', 'eigenvals'), optional. \cr}
}
\description{
Creates a pretty screeplpot using \code{\link[ggplot2]{ggplot}}. By default the explained variance is plotted
agaings the number of the principal component.
Alternatively the explained variance ratio, the cumulative
explained variance ratio, or the eigenvalues can be plotted.
}
\examples{
#

}
\seealso{
\code{\link[rsvd]{rpca}}, \code{\link[ggplot2]{ggplot}}
}
\author{
N. Benjamin Erichson, \email{erichson@berkeley.edu}
}
