% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rpca}
\alias{plot.rpca}
\title{Screeplot}
\usage{
\method{plot}{rpca}(x, ...)
}
\arguments{
\item{x}{Object returned by the \code{\link[rsvd]{rpca}} function.}

\item{...}{Additional arguments passed to the individual plot functions (see below).}
}
\description{
Creates a screeplot, variables and individual factor maps to
             summarize the results of the \code{\link[rsvd]{rpca}} function.
}
\examples{
#

}
\seealso{
\code{\link{ggscreeplot}}, \code{\link{ggcorplot}} , \code{\link{ggindplot}}
}
